Further refactoring to harvesters, transformations, and aggregations
