--- ecco_pipeline/field.py
@@ -1,11 +1,0 @@
-from dataclasses import dataclass
-from typing import Iterable
-
-@dataclass
-class Field():
-    name: str
-    long_name: str
-    standard_name: str
-    units: str
-    pre_transformations: Iterable[str]
-    post_transformations: Iterable[str]
