+++ ecco_pipeline/harvesters/catds_harvester.py
@@ -5,9 +5,8 @@ from typing import Iterable
 
 import requests
 from harvesters.enumeration.catds_enumerator import search_catds, CATDSGranule
-from harvesters.granule import Granule
-from harvesters.harvester import Harvester
-from utils.file_utils import get_date
+from harvesters.harvesterclasses import Harvester, Granule
+from utils.pipeline_utils.file_utils import get_date
 
 logger = logging.getLogger('pipeline')
