+++ ecco_pipeline/harvesters/enumeration/catds_enumerator.py
@@ -5,7 +5,7 @@ import os
 from bs4 import BeautifulSoup
 from datetime import datetime
 
-from harvesters.harvester import Harvester
+from harvesters.harvesterclasses import Harvester
 
 logger = logging.getLogger('pipeline')
